Update g-trasym.ad{b,s} from GCC 4.7 to avoid a libaddr2line dep
Before this patch, any attempt to use the GNAT symbolic traceback
package (as done, e.g., by AdaControl) would result in gnatlink
attempting to link against libaddr2line. This library was the result
of an AdaCore-specific patch to binutils that exposed the core
functionality of the addr2line utility.

Given that binutils never applied the AdaCore patch, this patch
follows the decision of upstream GCC to ship a symbolic traceback
package that instead prints the raw traceback addresses, which users
can manually run through addr2line.